chore: set up Alembic migrations; use Numeric/Decimal for money

+++ backend/app/models/payment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, String, ForeignKey, func, DateTime
+from sqlalchemy import Column, Integer, String, ForeignKey, func, DateTime, Numeric
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -9,7 +9,7 @@ class Payment(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     customer_id = Column(Integer, ForeignKey("customers.id"), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12,2), nullable=False)
     note = Column(String, nullable=True)
     created_at = Column(DateTime, server_default=func.now())
