+++ backend/app/models/sale.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column, Integer, ForeignKey, String, Float, func, DateTime
+
+from sqlalchemy import Column, Integer, ForeignKey, String, func, DateTime, Numeric
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -11,13 +12,13 @@ class Sale(Base):
     customer_id = Column(Integer, ForeignKey("customers.id"), nullable=False)
 
     product_name = Column(String, nullable=False)
-    quantity_kg = Column(Float, nullable=False)
-    unit_price = Column(Float, nullable=False)
-    vat_rate = Column(Float, nullable=False, default=0)
+    quantity_kg = Column(Numeric(10,3), nullable=False)
+    unit_price = Column(Numeric(12,2), nullable=False)
+    vat_rate = Column(Numeric(5,2), nullable=False, default=0)
 
-    subtotal = Column(Float, nullable=False)
-    vat_amount = Column(Float, nullable=False)
-    total = Column(Float, nullable=False)
+    subtotal = Column(Numeric(12,2), nullable=False)
+    vat_amount = Column(Numeric(12,2), nullable=False)
+    total = Column(Numeric(12,2), nullable=False)
     created_at = Column(DateTime, server_default=func.now())
 
     customer = relationship("Customer", back_populates="sales")
